backend/requirements.txt
amqp
anyio
asgiref
async-timeout
attrs
autobahn
automat
billiard
certifi
cffi
channels
channels-redis
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudinary
constantly
cryptography
daphne
dateutils
django
django-cors-headers
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine
djangorestframework
djangorestframework-simplejwt
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
events
gunicorn
h11
hyperlink
idna
incremental
iniconfig
kombu
msgpack
msgpack-python
packaging
pillow
pip
pluggy
prompt-toolkit
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pygments
pyjwt
pyopenssl
pytest
pytest-django
python-dateutil
python-dotenv
pytz
redis
requests
service-identity
setuptools
six
sniffio
sqlparse
tblib
twisted
txaio
typing-extensions
tzdata
urllib3
uuid
uvicorn
uvicorn-worker
vine
wcwidth
websockets
wheel
whitenoise
zope-interface
